Update gitignore again again lol
Remove /run/ directory

Remove /run-data/ directory

Add Block of Nuclear Waste

Add client and server communication for particles

Add decontaminator particle effect (way better than original hbm)

Add Radiation Decontaminator functionality

Refactor and Recode capability system for cleanliness

Add beginnings of HazardBlock 3d space effects (touching block causing rad poisoning + being near it also does so)

Add Nuclear Creeper. (No mushroom cloud yet :sadge:)

Add all capability types and values to both players and mobs

Break damage sources (cuz I'm cool like that)

Next things to come:
Fixing damage sources lmao
Adding json generators for damage sources (and achievements?)

+++ src/main/java/com/hbm/nucleartech/modules/ItemHazardModule.java
@@ -107,7 +107,8 @@ public class ItemHazardModule {
             if(reacher)
                 rad = (double) Math.min(Math.sqrt(rad), rad);
 
-            ContaminationUtil.contaminate(entity, HazardType.RADIATION, ContaminationType.CREATIVE, rad);
+//            System.err.println("calling ContaminationUtil.contaminate() for " + entity.getName().getString() + " with rad value: " + rad);
+            ContaminationUtil.contaminate(entity, HazardType.RADIATION, ContaminationType.CREATIVE, (float) rad);
         }
     }
